phongtroapi/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-vi-address==0.1.6
djangorestframework==3.15.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2024.2
